feat: Enhance progress note rendering with section detection and styling
- Added a new module `note_components.py` for rendering clinical progress notes with section headings.
- Implemented CSS loading for progress notes and added color styles in `colors.css`.
- Updated `views.py` to utilize the new progress note rendering functionality.
- Modified `xml_utils.py` to include color CSS in the transformation process.
- Refactored `trend_components.py` to improve patient trend rendering.
- Updated GitHub Actions workflow to detect stylesheet changes and create pull requests only when changes are present.
- Cleaned up and organized CSS files for better maintainability.

Closes #17
