+++ .github/workflows/update-stylesheets.yml
@@ -19,14 +19,15 @@ jobs:
       - run: pip install -r requirements.txt
       - run: python .github/scripts/update_stylesheets.py
       - name: Detect stylesheet changes
+        id: detect_stylesheet_changes
         run: |
           git add -A
           if git diff --cached --quiet; then
-            echo "no_changes=true" >> "$GITHUB_ENV"
+            echo "has_changes=false" >> "$GITHUB_OUTPUT"
           else
-            echo "no_changes=false" >> "$GITHUB_ENV"
+            echo "has_changes=true" >> "$GITHUB_OUTPUT"
           fi
-      - if: env.no_changes == 'false'
+      - if: steps.detect_stylesheet_changes.outputs.has_changes == 'true'
         uses: peter-evans/create-pull-request@v6
         with:
           commit-message: 'chore: Update CDA stylesheets'
