+++ frontend/static/colors.css
@@ -1,0 +1,25 @@
+/* Custom CSS for CDA Documents */
+/*https://coolors.co/2b4162-385f71-f5f0f6-d7b377-8f754f*/
+:root {
+/* CSS HEX */
+    --indigo-dye: #2b4162ff;
+    --paynes-gray: #385f71ff;
+    --magnolia: #f5f0f6ff;
+    --ecru: #d7b377ff;
+    --chamoisee: #8f754fff;
+
+    
+    --primary-color: var(--indigo-dye);
+    --secondary-color: var(--paynes-gray);
+    --text-color: #000000;
+    --background-color: var(--magnolia);
+    --section-bg: #ffffff;
+    --border-color: var(--chamoisee);
+    --header-bg: var(--magnolia);
+    --table-header-bg: var(--paynes-gray);
+    --table-header-text: #ffffff;
+    --table-stripe-bg: #ffffff;
+    --timestamp-color:var(--indigo-dye);
+    --footer-color: var(--chamoisee);
+    --clinical-border: var(--secondary-color);
+}
